feat: Update go version.
Note: I know some people have asked noot to depend on the sort.Slices, which is still flagged as
experimental.  However, it won't be one day it is well implemented. It is only used in one
place in the runtime and two places in an optional build tag antlr.statistics, I advise
users to make a fork if the really need to remove sort.Slice.

Signed-off-by: Jim.Idle <jimi@idle.ws>